services/requirements_test.txt
pytest==4.3.0
pytest-django==3.4.8
